NIFI-9625 Added check for cache directory existence before clean

+++ nifi-nar-bundles/nifi-standard-services/nifi-distributed-cache-services-bundle/nifi-distributed-cache-server/src/test/java/org/apache/nifi/distributed/cache/server/map/TestDistributedMapServerAndClient.java
@@ -67,7 +67,9 @@ public class TestDistributedMapServerAndClient {
 
     @BeforeEach
     public void setRunner() throws InitializationException, IOException {
-        FileUtils.cleanDirectory(dataFile);
+        if (dataFile.exists()) {
+            FileUtils.cleanDirectory(dataFile);
+        }
 
         runner = TestRunners.newTestRunner(Mockito.mock(Processor.class));
