+++ nifi-nar-bundles/nifi-standard-services/nifi-distributed-cache-services-bundle/nifi-distributed-cache-server/src/test/java/org/apache/nifi/distributed/cache/server/set/TestDistributedSetServerAndClient.java
@@ -60,7 +60,9 @@ public class TestDistributedSetServerAndClient {
 
     @BeforeEach
     public void setRunner() throws InitializationException, IOException {
-        FileUtils.cleanDirectory(dataFile);
+        if (dataFile.exists()) {
+            FileUtils.cleanDirectory(dataFile);
+        }
 
         runner = TestRunners.newTestRunner(Mockito.mock(Processor.class));
